+++ packages/validator/src/permission-guard/models/extension.ts
@@ -25,6 +25,7 @@ export interface ExtensionModule {
   isEnabled: boolean
   subModules: ExtensionSubModule[]
   permission?: RolePermission
+  allowedNamespaces?: string[]
 }
 
 export interface ExtensionManifest {